canvas-cli/requirements.txt
certifi==2020.11.8
chardet==3.0.4
click==7.1.2
idna==2.10
python-dateutil==2.8.1
requests==2.25.0
six==1.15.0
urllib3==1.26.2
